scanrouter-daemon: use scanrouter.conf, allow manual processing of scan dir
The daemon now uses scanrouter.conf and does basic command line parameter
parsing. The receiver processing was slightly abstracted.

For easy debugging, the daemon can be started with a directory as command line
parameter. The directory is assumed to be a scan job directory and is processed
accordingly. The daemon exits after having processed the directory.