scanrouter-daemon: fix empty_split mode, invalid num_empty_pages parsing
The empty_split mode was completely broken, as it didn't split the document
after num_empty_pages. This was caused by reading num_empty_pages as string
instead of as integer from the configuration file.

The patch correctly parses the configuration file, displays the num_empty_pages
value in the debug log message and saves the day.